+++ bindings/core/src/method/secret_manager.rs
@@ -16,6 +16,7 @@ use crate::OmittedDebug;
 #[derive(Clone, Derivative, Serialize, Deserialize)]
 #[derivative(Debug)]
 #[serde(tag = "name", content = "data", rename_all = "camelCase")]
+#[non_exhaustive]
 pub enum SecretManagerMethod {
     /// Generate Ed25519 addresses.
     GenerateEd25519Addresses {