+++ sdk/examples/client/output/all.rs
@@ -25,7 +25,7 @@ use iota_sdk::{
             AliasId, AliasOutputBuilder, BasicOutputBuilder, FoundryId, FoundryOutputBuilder, NativeToken, NftId,
             NftOutputBuilder, Output, OutputId, SimpleTokenScheme, TokenId, TokenScheme,
         },
-        payload::{transaction::TransactionEssence, Payload},
+        payload::Payload,
     },
 };
 
@@ -237,8 +237,7 @@ async fn main() -> Result<()> {
 fn get_alias_output_id(payload: &Payload) -> Result<OutputId> {
     match payload {
         Payload::Transaction(tx_payload) => {
-            let TransactionEssence::Regular(regular) = tx_payload.essence();
-            for (index, output) in regular.outputs().iter().enumerate() {
+            for (index, output) in tx_payload.essence().as_regular().outputs().iter().enumerate() {
                 if let Output::Alias(_alias_output) = output {
                     return Ok(OutputId::new(tx_payload.id(), index.try_into().unwrap())?);
                 }
@@ -253,8 +252,7 @@ fn get_alias_output_id(payload: &Payload) -> Result<OutputId> {
 fn get_foundry_output_id(payload: &Payload) -> Result<OutputId> {
     match payload {
         Payload::Transaction(tx_payload) => {
-            let TransactionEssence::Regular(regular) = tx_payload.essence();
-            for (index, output) in regular.outputs().iter().enumerate() {
+            for (index, output) in tx_payload.essence().as_regular().outputs().iter().enumerate() {
                 if let Output::Foundry(_foundry_output) = output {
                     return Ok(OutputId::new(tx_payload.id(), index.try_into().unwrap())?);
                 }
@@ -269,8 +267,7 @@ fn get_foundry_output_id(payload: &Payload) -> Result<OutputId> {
 fn get_nft_output_id(payload: &Payload) -> Result<OutputId> {
     match payload {
         Payload::Transaction(tx_payload) => {
-            let TransactionEssence::Regular(regular) = tx_payload.essence();
-            for (index, output) in regular.outputs().iter().enumerate() {
+            for (index, output) in tx_payload.essence().as_regular().outputs().iter().enumerate() {
                 if let Output::Nft(_nft_output) = output {
                     return Ok(OutputId::new(tx_payload.id(), index.try_into().unwrap())?);
                 }