+++ sdk/src/wallet/account/types/mod.rs
@@ -267,6 +267,7 @@ impl Transaction {
 
 /// Possible InclusionStates for transactions
 #[derive(Clone, Copy, Debug, Eq, Hash, PartialEq, Serialize, Deserialize)]
+#[non_exhaustive]
 pub enum InclusionState {
     Pending,
     Confirmed,
@@ -308,6 +309,7 @@ impl FromStr for OutputKind {
 /// The account identifier.
 #[derive(Debug, Clone, Serialize, Eq, PartialEq, Hash)]
 #[serde(untagged)]
+#[non_exhaustive]
 pub enum AccountIdentifier {
     /// Account alias as identifier.
     Alias(String),